shared/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["AshisGhosh <ashisghosh@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
litellm = "^1.34.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
